% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoimport-package.R
\docType{package}
\name{autoimport-package}
\alias{autoimport-package}
\title{autoimport: Automatic Generation of @importFrom Tags}
\description{
A toolbox to read all R files inside a package and automatically generate @importFrom 'roxygen2' tags in the right place. Includes a shiny application to review the changes before applying them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DanChaltiel/autoimport}
  \item \url{https://danchaltiel.github.io/autoimport/}
  \item Report bugs at \url{https://github.com/DanChaltiel/autoimport/issues}
}

}
\author{
\strong{Maintainer}: Dan Chaltiel \email{dan.chaltiel@gmail.com} (\href{https://orcid.org/0000-0003-3488-779X}{ORCID})

}
\keyword{internal}
