% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoimport.R
\name{autoimport}
\alias{autoimport}
\title{Automatically add \verb{@importFrom} tags}
\usage{
autoimport(
  files = dir("R/", full.names = TRUE),
  pkg_name = get_package_name(),
  namespace_file = "./NAMESPACE",
  description_file = "./DESCRIPTION",
  use_cache = TRUE,
  ask = TRUE,
  ignore_package = TRUE,
  verbose = 2
)
}
\arguments{
\item{files}{files to read. Default to the \verb{R/} folder.}

\item{pkg_name}{name of the package (character)}

\item{namespace_file}{path to the NAMESPACE file}

\item{description_file}{path to the DESCRIPTION file}

\item{use_cache}{use the cache system}

\item{ask}{whether to ask the user when multiple choices arise}

\item{ignore_package}{ignore files ending with \code{-package.R}}

\item{verbose}{the higher, the more output printed}
}
\value{
Nothing, used for side effects.
}
\description{
Automatically read all \code{R} files and add appropriate \verb{@importFrom} tags in the roxygen headers.
}
