% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{import_review}
\alias{import_review}
\alias{review_files}
\title{Decision management}
\source{
inspired by testthat::snapshot_review()
}
\usage{
import_review(
  data_files = review_files(),
  background = getOption("autoimport_background", FALSE)
)

review_files(source_path = "R/", output_path = get_target_dir())
}
\arguments{
\item{data_files}{result of \code{review_files()}}

\item{background}{whether to run the app in a background process. Default to \code{getOption("autoimport_background", FALSE)}.}

\item{path, output_path}{mostly used for tests}
}
\value{
nothing if \code{background==FALSE}, the (\link[callr:reexports]{callr::process}) object if \code{background==TRUE}
}
\description{
\code{import_review()} opens a Shiny app that shows a visual diff of each modified file. \cr
\code{review_files()} is a helper you probably won't ever use directly.
}
\section{Warning}{

Beware that using \code{background=TRUE} can bloat your system with multiple R session! \cr
You should probably kill the process when you are done:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{p=import_review(background=TRUE)
p$kill()
}\if{html}{\out{</div>}}
}

