% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importlist.R
\name{update_importlist}
\alias{update_importlist}
\alias{get_importlist}
\title{Update the \code{IMPORTLIST} file}
\usage{
update_importlist(imports, path = NULL)

get_importlist(path = NULL)
}
\arguments{
\item{imports}{a list of imports with \verb{key=function} and \code{value=package}}

\item{path}{path to the \code{IMPORTLIST} file}
}
\value{
nothing
}
\description{
Update the \code{IMPORTLIST} file, which forces the import of some packages without asking.
}
