% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoimport-package.R
\docType{package}
\name{autoimport-package}
\alias{autoimport-package}
\title{autoimport: Automatically generate @importFrom roxygen tags}
\description{
A toolbox to read all R files inside a package and automatically generate @importFrom roxygen in the right place.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DanChaltiel/autoimport}
  \item Report bugs at \url{https://github.com/DanChaltiel/autoimport/issues}
}

}
\author{
\strong{Maintainer}: Dan Chaltiel \email{dan.chaltiel@gmail.com} (\href{https://orcid.org/0000-0003-3488-779X}{ORCID})

}
\keyword{internal}
