% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoimport.R
\name{autoimport}
\alias{autoimport}
\title{Automatically compute \verb{@importFrom} tags}
\usage{
autoimport(
  root = ".",
  files = get_R_dir(root),
  namespace_file = "NAMESPACE",
  description_file = "DESCRIPTION",
  use_cache = TRUE,
  ignore_package = TRUE,
  verbose = 2
)
}
\arguments{
\item{root}{Path to the root of the package.}

\item{files}{Files to read. Default to the \verb{R/} folder.}

\item{namespace_file}{Path to the NAMESPACE file}

\item{description_file}{Path to the DESCRIPTION file}

\item{use_cache}{Whether to use the cache system. Can only be "read" or "write".}

\item{ignore_package}{Whether to ignore files ending with \code{-package.R}}

\item{verbose}{The higher, the more output printed. May slow the process a bit.}
}
\value{
Mostly used for side effects. Invisibly returns a dataframe summarizing the function imports, with input arguments as attributes.
}
\description{
Automatically read all \code{R} files and compute appropriate \verb{@importFrom} tags in the roxygen headers.
Choose which tags to add in the shiny app \code{\link[=import_review]{import_review()}} afterward.
}
