% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{import_review}
\alias{import_review}
\alias{review_files}
\alias{import_review_bg}
\title{Decision management}
\source{
shamelessly inspired by testthat::snapshot_review()
}
\usage{
import_review(data_files = review_files())

review_files(path = "R/")

import_review_bg(data_files = review_files())
}
\arguments{
\item{data_files}{result of \code{review_files()}}

\item{path}{mostly used for tests}
}
\description{
\itemize{
\item \code{import_review()} opens a Shiny app that shows a visual diff of each modified file.
\item \code{import_review_bg()} opens the Shiny app in a background job so that you can keep working. Disclaimer: this can bloat your stack call!
\item \code{review_files()} is a helper you probably won't need
}
}
